refactor : 패키지 구조 변경 및 toString()을 text()로 변경
