+++ src/main/java/wooteco/subway/line/dao/H2LineDao.java
@@ -1,10 +1,13 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dao;
 
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 import org.springframework.stereotype.Repository;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.domain.LineColor;
+import wooteco.subway.line.domain.LineName;
 
 import java.sql.PreparedStatement;
 import java.util.List;
@@ -32,8 +35,8 @@ public class H2LineDao implements LineDao {
         KeyHolder keyHolder = new GeneratedKeyHolder();
         jdbcTemplate.update(connection -> {
             PreparedStatement ps = connection.prepareStatement(sql, new String[]{"id"});
-            ps.setString(1, line.getName().toString());
-            ps.setString(2, line.getColor().toString());
+            ps.setString(1, line.getName().text());
+            ps.setString(2, line.getColor().text());
             return ps;
         }, keyHolder);
         long lineId = keyHolder.getKey().longValue();
@@ -74,7 +77,7 @@ public class H2LineDao implements LineDao {
                 "FROM LINE " +
                 "WHERE NAME = ?";
 
-        int countOfName = jdbcTemplate.queryForObject(sql, Integer.class, name.toString());
+        int countOfName = jdbcTemplate.queryForObject(sql, Integer.class, name.text());
         return countOfName > 0;
     }
 
@@ -84,7 +87,7 @@ public class H2LineDao implements LineDao {
                 "FROM LINE " +
                 "WHERE COLOR = ?";
 
-        int countOfColor = jdbcTemplate.queryForObject(sql, Integer.class, color.toString());
+        int countOfColor = jdbcTemplate.queryForObject(sql, Integer.class, color.text());
         return countOfColor > 0;
     }
 
@@ -104,7 +107,7 @@ public class H2LineDao implements LineDao {
                 "SET NAME = ?, COLOR = ? " +
                 "WHERE ID = ?";
 
-        jdbcTemplate.update(sql, line.getName().toString(), line.getColor().toString(), line.getId());
+        jdbcTemplate.update(sql, line.getName().text(), line.getColor().text(), line.getId());
     }
 
     @Override
