+++ src/main/java/wooteco/subway/line/dao/InMemoryLineDao.java
@@ -1,6 +1,9 @@
-package wooteco.subway.line;
+package wooteco.subway.line.dao;
 
 import org.springframework.util.ReflectionUtils;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.domain.LineColor;
+import wooteco.subway.line.domain.LineName;
 
 import java.lang.reflect.Field;
 import java.util.ArrayList;
