+++ src/main/java/wooteco/subway/line/domain/LineName.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.domain;
 
 import java.util.Objects;
 import java.util.regex.Pattern;
@@ -45,8 +45,7 @@ public class LineName {
         return Objects.hash(name);
     }
 
-    @Override
-    public String toString() {
+    public String text() {
         return name;
     }
 }
