+++ src/main/java/wooteco/subway/line/service/LineService.java
@@ -1,6 +1,8 @@
-package wooteco.subway.line;
+package wooteco.subway.line.service;
 
 import org.springframework.stereotype.Service;
+import wooteco.subway.line.domain.Line;
+import wooteco.subway.line.dao.LineDao;
 
 import java.util.List;
