+++ src/main/java/wooteco/subway/station/controller/StationController.java
@@ -1,8 +1,12 @@
-package wooteco.subway.station;
+package wooteco.subway.station.controller;
 
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import wooteco.subway.station.dto.StationRequest;
+import wooteco.subway.station.dto.StationResponse;
+import wooteco.subway.station.service.StationService;
+import wooteco.subway.station.domain.Station;
 
 import java.net.URI;
 import java.util.List;
@@ -22,7 +26,7 @@ public class StationController {
     public ResponseEntity<StationResponse> createStation(@RequestBody StationRequest stationRequest) {
         Station station = new Station(stationRequest.getName());
         Station newStation = stationService.save(station);
-        StationResponse stationResponse = new StationResponse(newStation.getId(), newStation.getName().toString());
+        StationResponse stationResponse = new StationResponse(newStation.getId(), newStation.getName().text());
         return ResponseEntity.created(URI.create("/stations/" + newStation.getId())).body(stationResponse);
     }
 
@@ -30,7 +34,7 @@ public class StationController {
     public ResponseEntity<List<StationResponse>> showStations() {
         List<Station> stations = stationService.findAll();
         List<StationResponse> stationResponses = stations.stream()
-                .map(station -> new StationResponse(station.getId(), station.getName().toString()))
+                .map(station -> new StationResponse(station.getId(), station.getName().text()))
                 .collect(Collectors.toList());
         return ResponseEntity.ok().body(stationResponses);
     }
