+++ src/main/java/wooteco/subway/station/dto/StationRequest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.station;
+package wooteco.subway.station.dto;
 
 public class StationRequest {
     private String name;
