+++ src/test/java/wooteco/subway/line/controller/LineAcceptanceTest.java
@@ -1,4 +1,4 @@
-package wooteco.subway.line;
+package wooteco.subway.line.controller;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -8,6 +8,8 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import wooteco.subway.AcceptanceTest;
+import wooteco.subway.line.dao.LineDao;
+import wooteco.subway.line.dto.LineResponse;
 
 import java.util.*;
 import java.util.stream.Collectors;
